fix: restore required Blueprint::getFromUser and correct companion migrations
Caught by running the migrations on a real Flarum 2.0.0-rc.1 install — all
three were invisible to php -l and code review.

- Blueprints: BlueprintInterface still REQUIRES getFromUser() in Flarum 2
  (getSender() is the extra, non-interface method). Removing getFromUser()
  in a prior cleanup made both registered blueprints abstract, so the whole
  app fatal-errored on boot. Restore it on NewPost and NewReply.
- Migration 000025: companion user_id was unsignedBigInteger but core
  users.id is INT UNSIGNED (Flarum uses increments(), not bigIncrements),
  so MySQL rejected the FK with errno 3780. Use unsignedInteger for user_id;
  group_id stays unsignedBigInteger to match social_groups.id.
- Migration 000026: Flarum injects the schema builder into migration
  closures, not a ConnectionInterface, so the ConnectionInterface type hint
  threw a TypeError. Take Builder $schema and derive the connection via
  $schema->getConnection().

Verified end-to-end: column dropped from users, companion table created with
both FKs (user_id->users.id, group_id->social_groups.id), and the existing
primary-group row backfilled.

+++ migrations/2024_01_01_000025_create_social_group_user_primary_table.php
@@ -20,7 +20,11 @@ return [
         }
 
         $schema->create('social_group_user_primary', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')->primary();
+            // user_id matches core users.id (INT UNSIGNED — Flarum uses
+            // increments(), not bigIncrements). group_id matches
+            // social_groups.id (BIGINT UNSIGNED). Mismatched integer widths
+            // make MySQL reject the foreign key with errno 3780.
+            $table->unsignedInteger('user_id')->primary();
             $table->unsignedBigInteger('group_id')->nullable()->index();
 
             $table->foreign('user_id')
